Moved some components to make it easier to read
- also fixed querying
